Leaving out --tags value should nonetheless default to all
This is a fix for optparse not handling an empty string value.

If you do not use --tags on the cmdline, it does default to all, which is good.  However, if you do specify --tags but leave out a value, it gets set to ¨¨, making sure nothing runs :(

Use case:
I have a Makefile that can take tags to only execute part of a playbook to speed things up and should execute all tasks when no tags are specified.
The ideal Makefile should simply have a   --tags=$(tags)   to handle both cases. My fix enables this.